# src/main/resources/static/doc/openapi.yaml
openapi: 3.0.1
info:
  title: Fastmoney API
  description: Esta API oferece operações bancárias seguras e convenientes, incluindo
    depósitos, saques, transferências e consulta de extrato. Os usuários precisam
    estar autenticados e têm permissão para apenas uma conta por CPF. As operações
    financeiras requerem um PIN válido. Há restrições, como saldo suficiente para
    saques e transferências, limites de R$ 500,00 para saques e transferências entre
    20h e 5h. A recuperação de senha é possível por email, e os usuários recebem notificações
    após cada operação. Esta API oferece uma solução completa e segura para as necessidades
    bancárias dos usuários.
  version: "0.0.1"
  contact:
    name: por email
    email: amrennan@gmail.com
  license:
    name: Licença MIT
    url: https://www.mit.edu/~amini/LICENSE.md
servers:
- url: "http://localhost:8080"
  description: Generated server url
paths:
  "/users":
    get:
      tags:
      - user-controller
      summary: Lista todos os usuários
      operationId: findAll
      parameters:
      - name: page
        in: query
        required: true
        schema:
          "$ref": "#/components/schemas/Pageable"
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/PageUserResponseDto"
      security:
        - bearer-key: []
    post:
      tags:
      - user-controller
      summary: Cadastra um novo usuário
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UserRequestDto"
        required: true
      responses:
        '201':
          description:  Created
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserResponseDto"
  "/users/{id}":
    get:
      tags:
      - user-controller
      summary: Busca usuário por id
      operationId: findById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserResponseDto"
      security:
        - bearer-key: []
    put:
      tags:
      - user-controller
      summary: Atualiza dados de um usuário
      operationId: update
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UserUpdateDto"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserResponseDto"
      security:
        - bearer-key: []          
    delete:
      tags:
      - user-controller
      summary: Exclui usuário
      operationId: delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '204':
          description: No Content
          content:
            "*/*":
              schema:
                type: string
      security:
        - bearer-key: []          
  "/login":
    post:
      tags:
      - auth-controller
      summary: Faz login na aplicação
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/LoginDTO"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
  "/transactions/{id}":
    get:
      tags:
      - transaction-controller
      summary: Extrato bancário
      operationId: statement
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: page
        in: query
        required: true
        schema:
          "$ref": "#/components/schemas/Pageable"
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/PageTransactionResponseDto"      
      security:
        - bearer-key: []          
  "/transactions/deposits/{id}":
    post:
      tags:
      - transaction-controller
      summary: Realiza depósitos
      operationId: deposit
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/TransactionRequestDto"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/TransactionResponseDto"              
      security:
        - bearer-key: []          
  "/transactions/withdrawals/{id}":
    post:
      tags:
      - transaction-controller
      summary: Realiza saques
      operationId: withdraw
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/TransactionRequestDto"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/TransactionResponseDto"
      security:
        - bearer-key: []          
  "/transactions/transfers/{senderId}/{receiverId}":
    post:
      tags:
      - transaction-controller
      summary: Realiza transferência entre usuários
      operationId: transfer
      parameters:
      - name: senderId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: receiverId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/TransactionRequestDto"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/TransactionResponseDto"
      security:
        - bearer-key: []          

components:
  schemas:
    UserUpdateDto:
      type: object
      properties:
        name:
          type: string
        pin:
          type: string
    AccountResponseDto:
      type: object
      properties:
        branch:
          type: string
        number:
          type: integer
          format: int32
        balance:
          type: number
    UserResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        cpf:
          type: string
        email:
          type: string
        accountDto:
          "$ref": "#/components/schemas/AccountResponseDto"
        active:
          type: boolean
    UserRequestDto:
      required:
      - accountBranch
      - email
      - name
      - password
      - pin
      type: object
      properties:
        name:
          type: string
        cpf:
          type: string
        email:
          type: string
        password:
          type: string
        accountBranch:
          type: string
        pin:
          type: string
    TransactionRequestDto:
      required:
      - pin
      - value
      type: object
      properties:
        value:
          type: number
        pin:
          type: string
    TransactionResponseDto:
      type: object
      properties:
        financialTransaction:
          type: string
          enum:
          - DEPOSIT
          - WITHDRAWAL
          - TRANSFER
        transactionType:
          type: string
          enum:
          - INPUT
          - OUTPUT
        date:
          type: string
          format: date-time
        value:
          type: number
        totalAfterOperation:
          type: number
        userDto:
          "$ref": "#/components/schemas/UserTransactionDto"
        receiverDto:
          "$ref": "#/components/schemas/UserTransactionDto"
    UserTransactionDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        accountBranch:
          type: string
        accountNumber:
          type: integer
          format: int32
    LoginDTO:
      type: object
      properties:
        login:
          type: string
        password:
          type: string
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
    PageUserResponseDto:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            "$ref": "#/components/schemas/UserResponseDto"
        number:
          type: integer
          format: int32
        sort:
          "$ref": "#/components/schemas/SortObject"
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          "$ref": "#/components/schemas/PageableObject"
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          "$ref": "#/components/schemas/SortObject"
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        paged:
          type: boolean
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
    PageTransactionResponseDto:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            "$ref": "#/components/schemas/TransactionResponseDto"
        number:
          type: integer
          format: int32
        sort:
          "$ref": "#/components/schemas/SortObject"
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          "$ref": "#/components/schemas/PageableObject"
        empty:
          type: boolean
  securitySchemes:
    bearer-key:
      type: http
      scheme: bearer
      bearerFormat: JWT